update all status for PR
Fixed all stations in red,
Fixed provider-url
Fixed file in wrong folder
Fixed station-names only with first letter in upper-case
Format to black.
